fix: id 통합

+++ app/src/main/java/com/hgh/na_o_man/data/dto/photo/response/PhotoUploadCountDto.kt
@@ -3,7 +3,7 @@ package com.hgh.na_o_man.data.dto.photo.response
 import com.hgh.na_o_man.domain.model.photo.PhotoUploadCountModel
 
 data class PhotoUploadCountDto(
-    val shareGroupId: Int,
+    val shareGroupId: Long,
     val uploadCount: Int
 ) {
     fun toModel() = PhotoUploadCountModel(
