+++ app/src/main/java/com/hgh/na_o_man/domain/model/share_group/GroupAddModel.kt
@@ -7,6 +7,6 @@ data class GroupAddModel(
     val memberCount: Int,
     val name: String,
     val profileInfoList: List<ProfileInfoModel>, //= emptyList(),
-    val shareGroupId: Int,
+    val shareGroupId: Long,
 
     )
