+++ app/src/main/java/com/hgh/na_o_man/domain/model/share_group/ProfileInfoModel.kt
@@ -2,7 +2,7 @@ package com.hgh.na_o_man.domain.model.share_group
 
 data class ProfileInfoModel(
     val image: String?,
-    val memberId: Int,
+    val memberId: Long,
     val name: String,
-    val profileId: Int
+    val profileId: Long
 )
